Update scraper.py
In this version, you'll see that we've defined two classes: WebScraper and GUI. WebScraper has methods for scraping a website and extracting data, as well as managing a queue of URLs to scrape. GUI manages the user interface, including a button to start scraping, an entry box to input the URL, and a text box to display the scraped data.